license debug
